Rename components. Add localization.
