+++ client/src/components/Layouts/App.js
@@ -6,7 +6,7 @@ import Dashboard from '../Dashboard/Page'
 import Landing from './Landing'
 import Signup from '../Signup/Page'
 import Login from '../Login/Page'
-import Projects from '../Projects/Projects'
+import Projects from '../Projects/Page'
 import Sales from '../Sales/Page'
 import SaleFormPage from '../Sales/FormPage'
 import requireAuth from '../../utils/requireAuth'
